+++ Sources/LBPresenter/NavReducer.swift
@@ -21,18 +21,3 @@ public struct NavReducer<NavState: Actionnable>: Sendable {
         self.navReducer(&navState, action)
     }
 }
-
-public struct Reducer<State: Actionnable, NavState: Actionnable>: Sendable {
-    let reducer: @Sendable (_ state: inout State, _ action: State.Action) -> Effect<State.Action, NavState.Action>
-
-    public init(reduce: @escaping @Sendable (_ state: inout State, _ action: State.Action) -> Effect<State.Action, NavState.Action>) {
-        self.reducer = reduce
-    }
-
-    /// Sends an action back into the system from an effect.
-    ///
-    /// - Parameter action: An action.
-    func callAsFunction(_ state: inout State, _ action: State.Action) -> Effect<State.Action, NavState.Action> {
-        self.reducer(&state, action)
-    }
-}
