+++ Sources/LBPresenter/PresenterState.swift
@@ -5,15 +5,6 @@
 //  Created by Rémi Lanteri on 25/11/2024.
 //
 
-public protocol Actionnable {
-    /// The type of actions that can be handled by the presenter.
-    ///
-    /// Actions represent events, user interactions, or commands that trigger
-    /// changes to the state or produce side effects. Conforming types must ensure
-    /// `Action` is `Sendable` to enable safe usage in asynchronous contexts.
-    associatedtype Action: Sendable
-}
-
 /// A protocol that defines the requirements for a state object used in a `Presenter`,
 /// enabling consistent management of UI state and actions.
 ///
