+++ Sources/LBPresenter/Utils/CancellableCollection.swift
